+++ web/src/refresh-pages/admin/ChatPreferencesPage.tsx
@@ -379,8 +379,8 @@ export default function ChatPreferencesPage() {
   const settings = useSettingsContext();
   const s = settings.settings;
   // Search Mode toggle is Business+; Chat Retention is Enterprise-only.
-  const canToggleSearchMode = useTierAtLeast(Tier.BUSINESS);
-  const canChangeChatRetention = useTierAtLeast(Tier.ENTERPRISE);
+  const businessTier = useTierAtLeast(Tier.BUSINESS);
+  const enterpriseTier = useTierAtLeast(Tier.ENTERPRISE);
 
   // Local state for text fields (save-on-blur)
   const [companyName, setCompanyName] = useState(s.company_name ?? "");
@@ -564,10 +564,10 @@ export default function ChatPreferencesPage() {
           <Card border="solid" rounding="lg">
             <Section>
               <Disabled
-                disabled={!canToggleSearchMode || uniqueSources.length === 0}
-                allowClick={canToggleSearchMode}
+                disabled={!businessTier || uniqueSources.length === 0}
+                allowClick={businessTier}
                 tooltip={
-                  !canToggleSearchMode
+                  !businessTier
                     ? "Search Mode requires the Business or Enterprise plan."
                     : "Set up connectors to use Search Mode"
                 }
@@ -575,7 +575,7 @@ export default function ChatPreferencesPage() {
                 <InputHorizontal
                   title="Search Mode"
                   tag={
-                    !canToggleSearchMode
+                    !businessTier
                       ? {
                           title: "Business Plan",
                           color: "amber",
@@ -584,17 +584,17 @@ export default function ChatPreferencesPage() {
                       : { title: "beta", color: "blue" }
                   }
                   description="UI mode for quick document search across your organization."
-                  disabled={!canToggleSearchMode || uniqueSources.length === 0}
+                  disabled={!businessTier || uniqueSources.length === 0}
                   withLabel
                 >
                   <Switch
                     checked={
-                      canToggleSearchMode ? s.search_ui_enabled ?? true : false
+                      businessTier ? s.search_ui_enabled ?? true : false
                     }
                     onCheckedChange={(checked) => {
                       void saveSettings({ search_ui_enabled: checked });
                     }}
-                    disabled={!canToggleSearchMode || uniqueSources.length === 0}
+                    disabled={!businessTier || uniqueSources.length === 0}
                   />
                 </InputHorizontal>
               </Disabled>
@@ -939,14 +939,14 @@ export default function ChatPreferencesPage() {
                 <Card border="solid" rounding="lg">
                   <Section>
                     <Disabled
-                      disabled={!canChangeChatRetention}
+                      disabled={!enterpriseTier}
                       tooltip="Chat history retention is an Enterprise Plan feature."
                     >
                       <InputHorizontal
                         title="Keep Chat History"
                         description="Specify how long Onyx should retain chats in your organization."
                         tag={
-                          !canChangeChatRetention
+                          !enterpriseTier
                             ? {
                                 title: "Enterprise Plan",
                                 color: "amber",
@@ -954,7 +954,7 @@ export default function ChatPreferencesPage() {
                               }
                             : undefined
                         }
-                        disabled={!canChangeChatRetention}
+                        disabled={!enterpriseTier}
                         withLabel
                       >
                         <InputSelect
@@ -970,7 +970,7 @@ export default function ChatPreferencesPage() {
                                   : parseInt(value, 10),
                             });
                           }}
-                          disabled={!canChangeChatRetention}
+                          disabled={!enterpriseTier}
                         >
                           <InputSelect.Trigger />
                           <InputSelect.Content>
